Compartmentalized ledge detection into own script
